python 0.0.3: Move RestClientSourceDatastore to datastore.py, i.e. support it
